requirements.txt
telethon>=1.28.5
gspread>=5.10.0
oauth2client>=4.1.3
python-dotenv>=1.0.0 